refactor animation code
